+++ dpa.py
@@ -29,9 +29,6 @@ def getUsefulTraceLength(fn):
   f.close()
   return c
 
-def loadTraces(fns):
-  dx = support.filemanager.load(fns)
-  return (dx['traces'],dx['data'])
 
 leakmodel = None
 
@@ -121,7 +118,6 @@ if __name__ == "__main__":
     sys.exit(0)
   print("Stage 1: Loading plaintexts...")
   tm = support.filemanager.TraceManager(fn)
-  # data,plaintexts = loadTraces(fn)
   print("Deriving key... wish me luck!")
   r = deriveKey(tm)
   if CONFIG_PLOT:
